リリースに向けたテストと修正が完了

+++ src/MxDBUnit/javasource/mxdbunit/implementation/ExcelDataLoader.java
@@ -3,19 +3,24 @@ package mxdbunit.implementation;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import org.apache.poi.openxml4j.opc.OPCPackage;
 import org.apache.poi.xssf.eventusermodel.XSSFReader;
 
 import com.mendix.core.Core;
+import com.mendix.logging.ILogNode;
 import com.mendix.systemwideinterfaces.core.IContext;
 import com.mendix.systemwideinterfaces.core.IMendixObject;
 
 import communitycommons.actions.deleteAll;
 
 public class ExcelDataLoader {
+	public static final ILogNode logger = Core.getLogger("MxDBUnit");
+
 	public static void loadAll(IContext context, String excelFilePath, String timeZoneId,
 			IdentityResolver identityResolver, boolean doClean)
 			throws Exception {
@@ -23,15 +28,24 @@ public class ExcelDataLoader {
 		final String[] currentEntityType = new String[1];
 		final List<IMendixObject> commitBuffer = new ArrayList<>();
 
-		String replacedFilePath = excelFilePath.replace("$HOME", System.getProperty("user.home")).replace("$RESOURCES",
-				Core.getConfiguration().getResourcesPath().getAbsolutePath());
+		String replacedFilePath = convertPath(excelFilePath);
 		File excelFile = new File(replacedFilePath);
 		if (doClean) {
+			Set<String> deletedEntities = (Set) context.getData().get("DeletedEntities");
+			if (deletedEntities == null) {
+				deletedEntities = new HashSet<>();
+				context.getData().put("DeletedEntities", deletedEntities);
+			}
 			List<String> targetEntities = getTargetEntitiesInOrder(excelFile);
 			// Perform deletion in reverse order (child to parent), taking dependencies into account.
 			Collections.reverse(targetEntities);
 			for (String entityType : targetEntities) {
+				if (deletedEntities.contains(entityType)) {
+					logger.warn("Entity {" + entityType + "} has already been deleted. Skipping deletion.");
+					continue; // Skip if already deleted
+				}
 				new deleteAll(context, entityType).executeAction();
+				deletedEntities.add(entityType);
 			}
 		}
 		XssfExcelReader.RowProcessor rowProcessor = new XssfExcelRowProcessor() {
@@ -42,7 +56,7 @@ public class ExcelDataLoader {
 					return;
 				}
 				IMendixObject obj = createIMendixObject(
-						context, currentEntityType[0], rowData, timeZoneId, identityResolver);
+						context, currentEntityType[0], rowData, timeZoneId, identityResolver, rowIndex);
 				commitBuffer.add(obj);
 				if (commitBuffer.size() >= 100) {
 					Core.commit(context, commitBuffer);
@@ -94,6 +108,14 @@ public class ExcelDataLoader {
 	}
 
 	private static boolean isTargetSheet(String sheetName) {
-		return !sheetName.startsWith("Expected_") && !sheetName.startsWith("#");
+		return !sheetName.startsWith("=") && !sheetName.startsWith("#");
+	}
+	
+	public static String convertPath(String path) {
+		String converted = path.replace("$MXDBUNIT", System.getenv("MXDBUNIT") != null ? System.getenv("MXDBUNIT") : "")
+				.replace("$HOME", System.getProperty("user.home"))
+				.replace("$RESOURCES", Core.getConfiguration().getResourcesPath().getAbsolutePath());
+		logger.debug("Converted excelFilePath from '" + path + "' to '" + converted + "'");
+		return converted;
 	}
 }
